+++ apps/dashboard/lib/main.dart
@@ -26,34 +26,11 @@ class DashboardApp extends StatefulWidget {
   State<DashboardApp> createState() => _DashboardAppState();
 }
 
-class _DashboardAppState extends State<DashboardApp>
-    with WidgetsBindingObserver {
+class _DashboardAppState extends State<DashboardApp> {
   late ThemeSettings _settings = widget.initialSettings;
 
   @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addObserver(this);
-  }
-
-  @override
-  void didChangePlatformBrightness() {
-    if (_settings.appearance == .system) setState(() {});
-  }
-
-  @override
-  void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    final brightness = switch (_settings.appearance) {
-      .light => Brightness.light,
-      .dark => Brightness.dark,
-      .system => WidgetsBinding.instance.platformDispatcher.platformBrightness,
-    };
     return ThemeScope(
       settings: _settings,
       onChanged: (settings) => setState(() => _settings = settings),
@@ -71,7 +48,7 @@ class _DashboardAppState extends State<DashboardApp>
           key: const ValueKey('dashboard-fortal-scope'),
           accent: _settings.accentColor,
           gray: _settings.grayColor,
-          brightness: brightness,
+          mode: _settings.appearance,
           panelBackground: _settings.panelBackground,
           radius: _settings.radius,
           scaling: _settings.scaling,