+++ apps/demo/lib/previews/preview_helper.dart
@@ -10,10 +10,7 @@ Widget createRemixPreview(Widget child) {
       settings: settings,
       pageBuilder: (context, animation, secondaryAnimation) => builder(context),
     ),
-    builder: (context, navigator) => FortalScope(
-      brightness: MediaQuery.platformBrightnessOf(context),
-      child: navigator!,
-    ),
+    builder: (context, navigator) => FortalScope(child: navigator!),
     home: ColoredBox(
       color: const Color(0xFFFAFAFA),
       child: Center(child: child),